rename refcards to blogs
