+++ src/content/blogs/06-exloring-nvim-config-with-docker.md
