+++ src/content/blogs/07-simplify-plugin-debugging-in-nvim.md
