+++ src/content/blogs/08-simplify-environment-variable-management-bash-script.md
