+++ src/layouts/Layout.astro
@@ -7,8 +7,8 @@ const navItems = [
   //	label: "Blog",
   //},
   {
-    path: "/refcards",
-    label: "Refcards",
+    path: "/blogs",
+    label: "Blogs",
   },
   // {
   // 	path: "/notes",
