+++ src/pages/blogs/script.js
